refactoring areas

+++ StarWarsTopTrumps.Engine/Hand.cs
@@ -2,8 +2,8 @@
 {
     public class Hand
     {
-        public long Player1Value { get; set; }
-        public long Player2Value { get; set; }
+        public decimal Player1Value { get; set; }
+        public decimal Player2Value { get; set; }
         public HandResult HandResult { get; set; }
         public string Message { get; set; }
     }
